# openapi.yaml
openapi: 3.0.0
info:
  title: Hostel Management API
  version: 1.0.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter your token in the text input below.
paths:
...
  /api/vacating-requests:
    get:
      summary: Get all vacating requests (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of vacating requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        tenant:
                          type: object
                          properties:
                            _id:
                              type: string
                            firstName:
                              type: string
                            lastName:
                              type: string
                            email:
                              type: string
                            room:
                              type: object
                              properties:
                                number:
                                  type: string
                        vacatingDate:
                          type: string
                          format: date
                        reason:
                          type: string
                        additionalNotes:
                          type: string
                        status:
                          type: string
                          enum: [pending, approved, rejected, completed]
                        approvedBy:
                          type: object
                          properties:
                            _id:
                              type: string
                            name:
                              type: string
                        approvedAt:
                          type: string
                          format: date-time
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
    post:
      summary: Create a new vacating request (tenant only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vacatingDate
                - reason
              properties:
                vacatingDate:
                  type: string
                  format: date
                  description: Date when tenant plans to vacate
                reason:
                  type: string
                  enum: [job-relocation, education-completed, personal-reasons, financial-issues, health-issues, found-alternative, other]
                  description: Reason for vacating
                additionalNotes:
                  type: string
                  description: Additional information or notes
      responses:
        '201':
          description: Vacating request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  request:
                    type: object
                    properties:
                      _id:
                        type: string
                      tenant:
                        type: string
                      vacatingDate:
                        type: string
                        format: date
                      reason:
                        type: string
                      additionalNotes:
                        type: string
                      status:
                        type: string
                        enum: [pending]
                      createdAt:
                        type: string
                        format: date-time
        '400':
          description: Bad request - Invalid data or duplicate request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Tenant access required
...
  /api/exchange-requests:
    get:
      summary: Get all exchange requests (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of exchange requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        tenant:
                          type: object
                          properties:
                            _id:
                              type: string
                            firstName:
                              type: string
                            lastName:
                              type: string
                            email:
                              type: string
                            room:
                              type: object
                              properties:
                                number:
                                  type: string
                        currentRoom:
                          type: string
                        desiredRoom:
                          type: string
                        reason:
                          type: string
                        preferredDate:
                          type: string
                          format: date
                        additionalNotes:
                          type: string
                        status:
                          type: string
                          enum: [pending, approved, rejected, completed]
                        approvedBy:
                          type: object
                          properties:
                            _id:
                              type: string
                            name:
                              type: string
                        approvedAt:
                          type: string
                          format: date-time
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
    post:
      summary: Create a new exchange request (tenant only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentRoom
                - desiredRoom
                - reason
                - preferredDate
              properties:
                currentRoom:
                  type: string
                  description: Current room number
                desiredRoom:
                  type: string
                  description: Desired room number
                reason:
                  type: string
                  enum: [roommate-compatibility, noise-issues, room-condition, location-preference, medical-reasons, academic-needs, other]
                  description: Reason for room exchange
                preferredDate:
                  type: string
                  format: date
                  description: Preferred exchange date
                additionalNotes:
                  type: string
                  description: Additional information or notes
      responses:
        '201':
          description: Exchange request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  request:
                    type: object
                    properties:
                      _id:
                        type: string
                      tenant:
                        type: string
                      currentRoom:
                        type: string
                      desiredRoom:
                        type: string
                      reason:
                        type: string
                      preferredDate:
                        type: string
                        format: date
                      additionalNotes:
                        type: string
                      status:
                        type: string
                        enum: [pending]
                      createdAt:
                        type: string
                        format: date-time
        '400':
          description: Bad request - Invalid data or room capacity issues
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Tenant access required
...